backend/requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels==3.0.1
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.3.0
docker==4.3.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pypiwin32==223
python-decouple==3.3
pytz==2020.1
pywin32==227
redis==3.5.3
requests==2.24.0
schedule==0.6.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted-20.3.0-cp39-cp39-win_amd64.whl
txaio==20.4.1
urllib3==1.25.11
websocket-client==0.57.0
zope.interface==5.2.0